SDG_Classifier_Streamlit_Application/requirements.txt
streamlit
regex
PyMuPDF
python-docx
transformers
torch
plotly
pandas
nltk